+++ src/pages/Member.js
@@ -1,7 +1,7 @@
 import React from 'react';
 
 import { useNavigate } from 'react-router-dom';
-import Header from './Header';
+import Header from '../components/Header';
 
 const Member = () => {
   const navigate = useNavigate(); // useNavigate를 호출
